Moved end_detect function to espnet_onnx.utils.function
